fix bugs with watch navigation
