+++ README.md
@@ -32,9 +32,9 @@ var movies = [
 
 
 ## Level 2 - Express Server
-- [ ] Move the hardcoded data from index.jsx to your Express Server js file
-- [ ] In your Express Server code a GET route called /movies which will be used to get the hardcoded list of movies
-- [ ] Modify your index.jsx to call this /movies GET route to do an initial load of your data
+- [x] Move the hardcoded data from index.jsx to your Express Server js file
+- [x] In your Express Server code a GET route called /movies which will be used to get the hardcoded list of movies
+- [x] Modify your index.jsx to call this /movies GET route to do an initial load of your data
 - [ ] In your Express Server code a POST route called /movie. This route will allow the user to create a new movie
 - [ ] Modify your index.jsx use the /movie POST route
